requirements.txt
PyPDF2==3.0.1
openai==0.27.6
google-generativeai==0.1.0rc1
sentence-transformers==2.3.0
langchain==0.0.284
python-dotenv==1.0.0
streamlit==1.22.0
faiss-cpu==1.7.4
tiktoken==0.4.0